seperated the footmu into smaller components
