+++ React_Practice/src/app/components/Body/ArtistPanel.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Pan_image from "./ArtistPanel/Pan_image"
+import Pan_image from "./ArtistPanel/Pan_image";
 import {Link} from "react-router";
 
 // link needs to be /Past_Projects/Band_stream/dist/index.html for live
@@ -9,7 +9,7 @@ export default class ArtistPanel extends React.Component {
     return (
       <div className="artist-panel col-xs-12 col-md-4">
         <div className="inner-artist-wrap col-xs-12">
-          <div className="mu-img-wrap col-xs-12"><Link to={"/Past_Projects/Band_stream/dist/index.html/songs"}>
+          <div className="mu-img-wrap col-xs-12"><Link to={"/songs"}>
             <Pan_image pan_image={this.props.pan_image}/>
           </Link></div>
           <div className="mu-info col-xs-12">
